Changed case statement.  Addded method for bad searches
I changed the "other" in the case statement to "all events".  This makes it easier and allows the user to see all the events on that day.

Also, added a wrong_search method to put after "else" within the case statement.  This eliminates all bad searches, whether a misspelled word or an integer search.